requirements.txt
# --- Shared & Server Dependencies ---
fastapi==0.136.0
uvicorn==0.44.0
httpx==0.28.1
pydantic==2.12.5
python-dotenv==1.0.1

# --- Middleware Dependencies (AI Providers) ---
# litellm handles all cloud AI providers internally.
# google-generativeai is deprecated (replaced by google-genai inside litellm).
# openai pinned to 1.x (stable SDK; litellm is compatible with both 1.x and 2.x).
litellm==1.67.5
openai>=1.13.0,<2.0.0
anthropic>=0.18.0,<1.0.0

# --- Backend Dependencies (AI & Processing) ---
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121
numpy>=1.26.0,<3.0.0
pandas>=2.0.0,<4.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
PyYAML>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
